renamed types to model folder
